+++ crates/kornia-apriltag/src/lib.rs
@@ -9,3 +9,6 @@ pub mod utils;
 
 /// Thresholding utilities for AprilTag detection.
 pub mod threshold;
+
+/// image iteration utilities module.
+pub mod iter;
